+++ vlib/builtin/cfns_wrapper.c.v
@@ -18,7 +18,7 @@ pub fn vstrlen_char(s &char) int {
 // The memory areas *MUST NOT OVERLAP*.  Use vmemmove, if the memory
 // areas do overlap. vmemcpy returns a pointer to `dest`.
 [inline; unsafe]
-pub fn vmemcpy(dest voidptr, const_src voidptr, n int) voidptr {
+pub fn vmemcpy(dest voidptr, const_src voidptr, n isize) voidptr {
 	unsafe {
 		return C.memcpy(dest, const_src, n)
 	}
@@ -30,7 +30,7 @@ pub fn vmemcpy(dest voidptr, const_src voidptr, n int) voidptr {
 // `src` or `dest`, and the bytes are then copied from the temporary array
 // to `dest`. vmemmove returns a pointer to `dest`.
 [inline; unsafe]
-pub fn vmemmove(dest voidptr, const_src voidptr, n int) voidptr {
+pub fn vmemmove(dest voidptr, const_src voidptr, n isize) voidptr {
 	unsafe {
 		return C.memmove(dest, const_src, n)
 	}
@@ -49,7 +49,7 @@ pub fn vmemmove(dest voidptr, const_src voidptr, n int) voidptr {
 // You should use a function that performs comparisons in constant time for
 // this.
 [inline; unsafe]
-pub fn vmemcmp(const_s1 voidptr, const_s2 voidptr, n int) int {
+pub fn vmemcmp(const_s1 voidptr, const_s2 voidptr, n isize) int {
 	unsafe {
 		return C.memcmp(const_s1, const_s2, n)
 	}
@@ -58,7 +58,7 @@ pub fn vmemcmp(const_s1 voidptr, const_s2 voidptr, n int) int {
 // vmemset fills the first `n` bytes of the memory area pointed to by `s`,
 // with the constant byte `c`. It returns a pointer to the memory area `s`.
 [inline; unsafe]
-pub fn vmemset(s voidptr, c int, n int) voidptr {
+pub fn vmemset(s voidptr, c int, n isize) voidptr {
 	unsafe {
 		return C.memset(s, c, n)
 	}
